python/rag-orchestrator/pyproject.toml
[project]
name = "rag-orchestrator"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "httpx>=0.28.0",
  "psycopg2-binary>=2.9.0",
  "pydantic>=2.11.0",
  "python-dotenv>=1.0.0",
  "uvicorn>=0.35.0",
  "rank-bm25>=0.2.2",
  "flashrank>=0.2.0",
]

[project.optional-dependencies]
eval = [
  "ragas>=0.2.0",
  "datasets>=3.0.0",
]

[project.scripts]
rag-orchestrator = "rag_orchestrator.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_orchestrator"]

[tool.uv]
dev-dependencies = []
